+++ plugin-name/public/class-plugin-name-public.php
@@ -31,6 +31,15 @@ class Plugin_Name_Public {
 	private $plugin_name;
 
 	/**
+	 * The unique prefix of this plugin.
+	 *
+	 * @since    1.0.0
+	 * @access   private
+	 * @var      string    $plugin_prefix    The string used to uniquely prefix technical functions of this plugin.
+	 */
+	private $plugin_prefix;
+
+	/**
 	 * The version of this plugin.
 	 *
 	 * @since    1.0.0
@@ -44,11 +53,13 @@ class Plugin_Name_Public {
 	 *
 	 * @since    1.0.0
 	 * @param      string $plugin_name      The name of the plugin.
+	 * @param      string $plugin_prefix          The unique prefix of this plugin.
 	 * @param      string $version          The version of this plugin.
 	 */
-	public function __construct( $plugin_name, $version ) {
+	public function __construct( $plugin_name, $plugin_prefix, $version ) {
 
-		$this->plugin_name = $plugin_name;
+		$this->plugin_name   = $plugin_name;
+		$this->plugin_prefix = $plugin_prefix;
 		$this->version = $version;
 
 	}